+++ lib/image_blur.dart
@@ -2,10 +2,11 @@ import 'dart:ui';
 
 import 'package:fast_cached_network_image/fast_cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:image_blur/src/tools/get_image_tools.dart';
+import 'package:image_blur/src/widget/image_blur_get_pallette_color.dart';
 import 'package:image_blur/src/widget/image_circular_blur.dart';
 import 'package:image_blur/src/widget/image_hash_preview.dart';
 import 'package:image_blur/src/widget/image_hash_preview_circular.dart';
+import 'package:image_blur/src/widget/image_hash_preview_get_pallette_color.dart';
 import 'package:image_blur/src/widget/image_hash_web.dart';
 import 'package:palette_generator/palette_generator.dart';
 
@@ -161,10 +162,6 @@ class ImageBlur extends StatefulWidget {
   ///  Be aware that clipping may be costly in terms of performance.
   final BorderRadiusGeometry borderRadius;
 
-  /// Fetches the image and generates the palette color
-  final Future<PaletteGenerator?>? Function(Future<PaletteGenerator>?)?
-      onPaletteReceived;
-
   const ImageBlur({
     required this.imageUrl,
     Key? key,
@@ -196,7 +193,6 @@ class ImageBlur extends StatefulWidget {
     this.memCacheWidth,
     this.placeholderColor = const Color.fromRGBO(224, 224, 224, 1),
     this.borderRadius = BorderRadius.zero,
-    this.onPaletteReceived,
   });
 
   static Widget imageCircularBlur({
@@ -602,9 +598,7 @@ class ImageBlur extends StatefulWidget {
     /// [borderRadius] The border radius of the image widget.
     final BorderRadiusGeometry borderRadius = BorderRadius.zero,
 
-    /// Fetches the image and generates the palette color
-    final Future<PaletteGenerator?>? Function(Future<PaletteGenerator>?)?
-        onPaletteReceived,
+    /// [key] The key for the image widget.
     Key? key,
   }) {
     return ImageHashPreview(
@@ -640,7 +634,6 @@ class ImageBlur extends StatefulWidget {
       cacheHeight: cacheHeight,
       scale: scale,
       borderRadius: borderRadius,
-      onPaletteReceived: onPaletteReceived,
       key: key,
     );
   }
@@ -851,6 +844,338 @@ class ImageBlur extends StatefulWidget {
     );
   }
 
+  static Widget imageBlurGetPalletteColor({
+    /// [imageUrl] The path or URL of the image to be displayed.
+    required final String imageUrl,
+
+    ///Default [BoxFit.cover]
+    final BoxFit? fit = BoxFit.cover,
+
+    /// [height] non-null, requires the child to have exactly this height.
+    final double? height,
+
+    /// [width] non-null, requires the child to have exactly this width.
+    final double? width,
+
+    ///Used to combine [color] with this image.
+    ///The default is [BlendMode.srcIn]. In terms of the blend mode, [color] is the source and this image is the destination.
+    ///See also:
+    ///[BlendMode], which includes an illustration of the effect of each blend mode.
+    final BlendMode? colorBlendMode = BlendMode.srcIn,
+
+    ///If non-null, this color is blended with each image pixel using [colorBlendMode].
+    final Color? color,
+
+    /// How to align the image within its bounds.
+
+    ///The alignment aligns the given position in the image to the given position in the layout bounds. For example, an [Alignment] alignment of (-1.0, -1.0) aligns the image to the top-left corner of its layout bounds, while an [Alignment] alignment of (1.0, 1.0) aligns the bottom right of the image with the bottom right corner of its layout bounds. Similarly, an alignment of (0.0, 1.0) aligns the bottom middle of the image with the middle of the bottom edge of its layout bounds.
+
+    ///To display a subpart of an image, consider using a [CustomPainter] and [Canvas.drawImageRect].
+
+    ///If the [alignment] is [TextDirection]-dependent (i.e. if it is a [AlignmentDirectional]), then an ambient [Directionality] widget must be in scope.
+
+    ///Defaults to [Alignment.center].
+
+    //See also:
+
+    ///[Alignment], a class with convenient constants typically used to specify an [AlignmentGeometry].
+    ///[AlignmentDirectional], like [Alignment] for specifying alignments relative to text direction.
+    final Alignment alignment = Alignment.center,
+
+    ///The center slice for a nine-patch image.
+    final Rect? centerSlice,
+
+    /// If non-null, the value from the [Animation] is multiplied with the opacity
+    /// of each image pixel before painting onto the canvas.
+    ///
+    /// This is more efficient than using [FadeTransition] to change the opacity
+    /// of an image, since this avoids creating a new composited layer. Composited
+    /// layers may double memory usage as the image is painted onto an offscreen
+    /// render target.
+    ///
+    /// See also:
+    ///
+    ///  * [AlwaysStoppedAnimation], which allows you to create an [Animation]
+    ///    from a single opacity value.
+    final Animation<double>? opacity,
+
+    /// The rendering quality of the image.
+    ///
+    /// {@template flutter.widgets.image.filterQuality}
+    /// If the image is of a high quality and its pixels are perfectly aligned
+    /// with the physical screen pixels, extra quality enhancement may not be
+    /// necessary. If so, then [FilterQuality.none] would be the most efficient.
+    ///
+    /// If the pixels are not perfectly aligned with the screen pixels, or if the
+    /// image itself is of a low quality, [FilterQuality.none] may produce
+    /// undesirable artifacts. Consider using other [FilterQuality] values to
+    /// improve the rendered image quality in this case. Pixels may be misaligned
+    /// with the screen pixels as a result of transforms or scaling.
+    ///
+    /// See also:
+    ///
+    ///  * [FilterQuality], the enum containing all possible filter quality
+    ///    options.
+    /// {@endtemplate}
+    final FilterQuality filterQuality = FilterQuality.low,
+
+    ///How to paint any portions of the layout bounds not covered by the image.
+    final ImageRepeat repeat = ImageRepeat.noRepeat,
+
+    /// Whether to paint the image in the direction of the [TextDirection].
+    ///
+    /// If this is true, then in [TextDirection.ltr] contexts, the image will be
+    /// drawn with its origin in the top left (the "normal" painting direction for
+    /// images); and in [TextDirection.rtl] contexts, the image will be drawn with
+    /// a scaling factor of -1 in the horizontal direction so that the origin is
+    /// in the top right.
+    ///
+    /// This is occasionally used with images in right-to-left environments, for
+    /// images that were designed for left-to-right locales. Be careful, when
+    /// using this, to not flip images with integral shadows, text, or other
+    /// effects that will look incorrect when flipped.
+    ///
+    /// If this is true, there must be an ambient [Directionality] widget in
+    /// scope.
+    final bool matchTextDirection = false,
+
+    /// Whether the image should be played in the background when it's not visible.
+    final bool gapLessPlayback = false,
+
+    /// The semantic label for this image.
+    final String? semanticLabel,
+
+    /// A builder function that is called if an error occurs during image loading.
+    final ImageFrameBuilder? frameBuilder,
+
+    /// A builder function that is called if an error occurs during image loading.
+    final ImageLoadingBuilder? loadingBuilder,
+
+    /// A builder function that is called if an error occurs during image loading.
+    final ImageErrorWidgetBuilder? errorBuilder,
+
+    /// Whether to paint the image with anti-aliasing.
+    final bool isAntiAlias = false,
+
+    /// The headers used for http requests.
+    final Map<String, String>? headers,
+
+    /// The width and height of the image in logical pixels.
+    final int? cacheWidth,
+
+    /// The width and height of the image in logical pixels.
+    final int? cacheHeight,
+
+    /// The color to use when drawing the image.
+    final TileMode tileMode = TileMode.decal,
+
+    /// The duration of the fade-in effect.
+    final Duration fadeInDuration = const Duration(milliseconds: 500),
+
+    ///The color of the placeholder image.
+    final Color backgroundImage = const Color.fromRGBO(238, 238, 238, 1),
+
+    /// The height and width of the image in logical pixels.
+    final double scale = 1.0,
+
+    /// The height and width of the image in logical pixels.
+    final int? memCacheHeight,
+
+    /// The height and width of the image in logical pixels.
+    final int? memCacheWidth,
+
+    ///The color of the placeholder image.
+    final Color? placeholderColor = const Color.fromRGBO(224, 224, 224, 1),
+
+    ///If non-null, the corners of this box are rounded by this [BorderRadius].
+    ///Applies only to boxes with rectangular shapes; ignored if [shape] is not [BoxShape.rectangle].
+    ///The [shape] or the [borderRadius] won't clip the children of the decorated [Container].
+    /// If the clip is required, insert a clip widget (e.g., [ClipRect], [ClipRRect], [ClipPath]) as the child of the [Container].
+    ///  Be aware that clipping may be costly in terms of performance.
+    final BorderRadiusGeometry borderRadius = BorderRadius.zero,
+
+    /// Fetches the image and generates the palette color
+    final Future<PaletteGenerator?>? Function(Future<PaletteGenerator>?)?
+        onPaletteReceived,
+
+    /// [key] for [ImageBlurGetPalletteColor]
+    Key? key,
+  }) {
+    return ImageBlurGetPalletteColor(
+      imageUrl: imageUrl,
+      height: height,
+      width: width,
+      borderRadius: borderRadius,
+      fit: fit,
+      onPaletteReceived: onPaletteReceived,
+      placeholderColor: placeholderColor,
+      memCacheHeight: memCacheHeight,
+      memCacheWidth: memCacheWidth,
+      filterQuality: filterQuality,
+      repeat: repeat,
+      matchTextDirection: matchTextDirection,
+      gapLessPlayback: gapLessPlayback,
+      semanticLabel: semanticLabel,
+      frameBuilder: frameBuilder,
+      loadingBuilder: loadingBuilder,
+      errorBuilder: errorBuilder,
+      isAntiAlias: isAntiAlias,
+      headers: headers,
+      cacheWidth: cacheWidth,
+      cacheHeight: cacheHeight,
+      tileMode: tileMode,
+      fadeInDuration: fadeInDuration,
+      backgroundImage: backgroundImage,
+      scale: scale,
+      alignment: alignment,
+      centerSlice: centerSlice,
+      color: color,
+      colorBlendMode: colorBlendMode,
+      opacity: opacity,
+      key: key,
+    );
+  }
+
+  static imageHashGetPaletteColor({
+    /// [imagePath] The path or URL of the image to be displayed.
+    required final String imagePath,
+
+    /// [width] The width of the image widget.
+    final double? width,
+
+    /// [height] The height of the image widget.
+    final double? height,
+
+    /// [placeholderColor] The color to display as a placeholder while the image is loading.
+    final Color? placeholderColor = const Color.fromRGBO(224, 224, 224, 1),
+
+    /// [curve] The curve used for animation transitions.
+    final Curve curve = Curves.easeOut,
+
+    /// [fit] How the image should be inscribed into the space allocated during layout.
+    final BoxFit fit = BoxFit.cover,
+
+    /// [decodingHeight] The height used to decode the image.
+    final int decodingHeight = 32,
+
+    /// [decodingWidth] The width used to decode the image.
+    final int decodingWidth = 32,
+
+    /// [duration] The duration of the transition animation.
+    final Duration duration = const Duration(milliseconds: 1000),
+
+    /// [onDecoded] A callback function invoked when the image is decoded.
+    final void Function()? onDecoded,
+
+    /// [onStarted] A callback function invoked when the image loading process starts.
+    final void Function()? onStarted,
+
+    /// [onReady] A callback function invoked when the image is ready for display.
+    final void Function()? onReady,
+
+    /// [onDisplayed] A callback function invoked when the image is displayed.
+    final void Function()? onDisplayed,
+
+    /// [colorBlendMode] The blend mode used to blend the image with the background color.
+    final BlendMode? colorBlendMode,
+
+    /// [color] The color applied as a filter to the image.
+    final Color? color,
+
+    /// [alignment] The alignment of the image within its bounding box.
+    final Alignment alignment = Alignment.center,
+
+    /// [centerSlice] The rectangle inside the image used for centering and scaling.
+    final Rect? centerSlice,
+
+    /// [opacity] The opacity of the image.
+    final Animation<double>? opacity,
+
+    /// [filterQuality] The quality of the image filtering.
+    final FilterQuality filterQuality = FilterQuality.low,
+
+    /// [repeat] The strategy to use when painting the image.
+    final ImageRepeat repeat = ImageRepeat.noRepeat,
+
+    /// [matchTextDirection] Whether to match the direction of the image with the direction of the text.
+    final bool matchTextDirection = false,
+
+    /// [gapLessPlayback] Whether to gaplessly loop a finite set of images.
+    final bool gapLessPlayback = false,
+
+    /// [semanticLabel] A semantic description of the image.
+    final String? semanticLabel,
+
+    /// [frameBuilder] A builder function used to create custom frames for the image.
+    final ImageFrameBuilder? frameBuilder,
+
+    /// [loadingBuilder] A builder function used to create custom widgets while the image is loading.
+    final ImageLoadingBuilder? loadingBuilder,
+
+    /// [errorBuilder] A builder function used to create custom error widgets.
+    final ImageErrorWidgetBuilder? errorBuilder,
+
+    /// [isAntiAlias] Whether to use anti-aliasing when painting the image.
+    final bool isAntiAlias = false,
+
+    /// [headers] Optional HTTP headers to include in the image request.
+    final Map<String, String>? headers,
+
+    /// [cacheWidth] The desired width of the image cache.
+    final int? cacheWidth,
+
+    /// [cacheHeight] The desired height of the image cache.
+    final int? cacheHeight,
+
+    /// [scale] The scale to apply to the image.
+    final double scale = 1.0,
+
+    /// [borderRadius] The border radius of the image widget.
+    final BorderRadiusGeometry borderRadius = BorderRadius.zero,
+
+    /// Fetches the image and generates the palette color
+    final Future<PaletteGenerator?>? Function(Future<PaletteGenerator>?)?
+        onPaletteReceived,
+    Key? key,
+  }) {
+    return ImageHashGetPaletteColor(
+      imagePath: imagePath,
+      width: width,
+      height: height,
+      placeholderColor: placeholderColor,
+      curve: curve,
+      fit: fit,
+      decodingHeight: decodingHeight,
+      decodingWidth: decodingWidth,
+      duration: duration,
+      onDecoded: onDecoded,
+      onReady: onReady,
+      onStarted: onStarted,
+      onDisplayed: onDisplayed,
+      colorBlendMode: colorBlendMode,
+      color: color,
+      alignment: alignment,
+      centerSlice: centerSlice,
+      opacity: opacity,
+      filterQuality: filterQuality,
+      repeat: repeat,
+      matchTextDirection: matchTextDirection,
+      gapLessPlayback: gapLessPlayback,
+      semanticLabel: semanticLabel,
+      frameBuilder: frameBuilder,
+      loadingBuilder: loadingBuilder,
+      errorBuilder: errorBuilder,
+      isAntiAlias: isAntiAlias,
+      headers: headers,
+      cacheWidth: cacheWidth,
+      cacheHeight: cacheHeight,
+      scale: scale,
+      borderRadius: borderRadius,
+      onPaletteReceived: onPaletteReceived,
+      key: key,
+    );
+  }
+
   @override
   State<ImageBlur> createState() => _ImageBlurState();
   static ImageBlur instance() => ImageBlur(
@@ -864,15 +1189,6 @@ class ImageBlur extends StatefulWidget {
 }
 
 class _ImageBlurState extends State<ImageBlur> {
-  @override
-  void initState() {
-    GetImage.fetchImageAndGeneratePalette(widget.imageUrl).then((value) {
-      widget.onPaletteReceived?.call(Future.value(value));
-      setState(() {});
-    });
-    super.initState();
-  }
-
   @override
   Widget build(BuildContext context) {
     double blurValue = 0.0;